+++ src/common/TodoBase.ts
@@ -1,4 +1,12 @@
+import { workspace } from 'vscode';
+
 export default abstract class TodoBase {
+  protected settings
+  constructor() {
+    this.settings = workspace.getConfiguration('todolens')
+  }
   public update(){}
-  public configChanged(){}
+  public configChanged(){
+    this.settings = workspace.getConfiguration('todolens')
+  }
 }