+++ src/extension.ts
@@ -1,67 +1,52 @@
-import {commands, languages, window, QuickPickItem, Selection, workspace, Uri, FileSystemError, Range, Position} from 'vscode'
-import * as _ from 'lodash'
+import {languages, window, QuickPickItem, Selection, workspace, Uri, FileSystemError, Range, Position} from 'vscode'
 import TodoLensProvider from './lens/TodoLensProvider'
 import Annotations from './common/Annotations'
 import Highlighter from './syntax/Highlighter'
+import Commands from './common/Commands'
+import TodoBase from './common/TodoBase'
 
-var lens, highlight, annotations, activeEditor, settings
+const modules : TodoBase[] = []
+
+var lens, highlight, commands, annotations, activeEditor, settings
 
 function activate(context) {
 	const disposables = []
 
 	settings = workspace.getConfiguration('todolens')
 	activeEditor = window.activeTextEditor
+
 	annotations = new Annotations()
+	modules.push(annotations)
 	highlight = new Highlighter(annotations)
+	modules.push(highlight)
 	lens = new TodoLensProvider(annotations)
+	modules.push(lens)
+	commands = new Commands(annotations)
+	modules.push(commands)
 
 	// list command
+	disposables.push(...commands.get())
+
+	// add the lens to the disposables
 	disposables.push(
-		commands.registerCommand("todolens.list", (args) => {
-			// get all the items
-			let arr : Array<QuickPickItem> = []
-			let an
-			if(args != undefined) {
-				an = args.split('|')
-			} else {
-				// get all annotations
-				an = annotations.getAll()
-			}
-			// filter out possible duplicates
-			an = _.uniqBy(an, (e) => {return e})
-			// get all the values by the given types
-			an.forEach(a => {
-				annotations.get(a).forEach(t => {
-					arr.push({
-						label: t.index + '',
-						description: t.text.trim()
-					})
-				})
-			})
-			let o = settings.get('dropdownOrdering', 'category')
-			switch(o) {
-				case 'line_numbers_asc': {
-					arr = _.orderBy(arr, [(r) => {return Number(r.label)}], ['asc'])
-					break
-				}
-				case 'line_numbers_desc': {
-					arr = _.orderBy(arr, [(r) => {return Number(r.label)}], ['desc'])
-					break
-				}
-			}
-			// show the quick picker
-			window.showQuickPick(arr, {canPickMany: false})
-			.then((v) => {
-				if(!v) return
-				let editor = window.activeTextEditor
-				let range = editor.document.lineAt(Number(v.label)-1).range
-				editor.selection =  new Selection(range.start, range.start)
-				editor.revealRange(new Range(range.start, new Position(range.end.line + (editor.visibleRanges[0].end.line - editor.visibleRanges[0].start.line) - 3, range.end.character)))
-			})
-		})
+		languages.registerCodeLensProvider({
+				language: "*",
+				scheme: "file"
+			},
+			lens
+		)
 	)
 
-	// events
+	setupEvents(context)
+	update()
+
+	context.subscriptions.push(...disposables)
+}
+
+/*
+ * Setup all the window / workspace events here
+ */
+function setupEvents(context) {
 	window.onDidChangeActiveTextEditor((e) => {
 		activeEditor = e
 		if(e) {
@@ -78,34 +63,31 @@ function activate(context) {
 	workspace.onDidChangeConfiguration(() => {
 		configChanged()
 	}, null, context.subscriptions)
-
-	disposables.push(
-		languages.registerCodeLensProvider({
-				language: "*",
-				scheme: "file"
-			},
-			lens
-		)
-	)
-
-	update()
-	context.subscriptions.push(...disposables)
 }
-exports.activate = activate
 
+/*
+ * Update all the modules
+ */
 function update() {
-	annotations.update()
-	if(settings.get('useHighlighting', true)) highlight.update()
+	modules.forEach(m => {
+		m.update()
+	})
 }
+
+/*
+ * Apply the config changes to the modules
+ */
 function configChanged() {
 	settings = workspace.getConfiguration('todolens')
-	// annotations.configChanged()
-	if(settings.get('useHighlighting', true)) highlight.configChanged()
+	modules.forEach(m => {
+		m.configChanged()
+	})
 	this.update()
 }
 
 function deactivate() {}
 
+exports.activate = activate
 module.exports = {
 	activate,
 	deactivate